Move main to modules
