+++ src/main/MainWindow.js
@@ -1,49 +1,45 @@
-'use strict';
-const { app, BrowserWindow, ipcMain, dialog, shell } = require('electron');
-
-const path = require('path');
-const os = require('os');
-
-const ConfigManager = require('./managers/ConfigManager');
-const { restoreWindow } = require('./helpers');
+const { app, BrowserWindow, shell } = require('electron');
 
 /**
  * @type {BrowserWindow} - The main window
  */
 var win = undefined;
+exports.window = win;
 
-exports.MainWindow = win;
+exports.create = () => new Promise((resolve, reject) => {
 
-const createMainWindow = () => new Promise((resolve, reject) => {
+	const path = require('path');
+	const ConfigManager = require('./managers/ConfigManager');
+	const Updater = require('./Updater');
 
 	require('./menu').createMenu();
 
 	let windowState = require('./libs/WindowState')({
-			width: 1280,
-			height: 720
+		width: 1280,
+		height: 720
 	});
 
 	win = new BrowserWindow({
-			x: windowState.x,
-			y: windowState.y,
-			width: windowState.width,
-			height: windowState.height,
-			minWidth: 720,
-			minHeight: 480,
-			show: false,
-			resizable: true,
-			frame: process.platform === 'darwin',
-			webPreferences: {
-					preload: path.join(__dirname, 'preloader.js'),
-					nodeIntegration: false,
-					contextIsolation: true,
-					worldSafeExecuteJavaScript: true,
-					spellcheck: true,
-					enableRemoteModule: true,
-			},
-			titleBarStyle: 'default',
-			roundedCorners: true,
-			backgroundColor: '#171614'
+		x: windowState.x,
+		y: windowState.y,
+		width: windowState.width,
+		height: windowState.height,
+		minWidth: 720,
+		minHeight: 480,
+		show: false,
+		resizable: true,
+		frame: process.platform === 'darwin',
+		webPreferences: {
+			preload: path.join(__dirname, 'preloader.js'),
+			nodeIntegration: false,
+			contextIsolation: true,
+			worldSafeExecuteJavaScript: true,
+			spellcheck: true,
+			enableRemoteModule: true,
+		},
+		titleBarStyle: 'default',
+		roundedCorners: true,
+		backgroundColor: '#171614'
 	});
 
 	require('@electron/remote/main').enable(win.webContents);
@@ -60,42 +56,47 @@ const createMainWindow = () => new Promise((resolve, reject) => {
 	win.on('focus', () => win.webContents.send('focus'));
 	win.on('closed', () => (win = null, app.quit()));
 	win.on('close', (event) => {
-			if (ConfigManager.getOption("launcher.hideOnClose")) {
-					event.preventDefault();
-					win.hide();
-			};
+		if (ConfigManager.getOption("launcher.hideOnClose")) {
+			event.preventDefault();
+			win.hide();
+		};
 	});
 
 	win.webContents.on("did-finish-load", () => {
-			console.debug("> finish load");
-			if (ConfigManager.getOption("launcher.checkUpdates")) {
-					// checkForUpdates();
-			}
+		console.debug("> finish load");
+		if (ConfigManager.getOption("launcher.checkUpdates")) {
+			Updater.checkForUpdates();
+		}
 	});
 
 	// handler for blank target
 	win.webContents.setWindowOpenHandler(({ url }) =>
-			url.startsWith("file://") ? { action: 'allow' } :
+		url.startsWith("file://") ? { action: 'allow' } :
 			(shell.openExternal(url), { action: 'deny' })
 	);
 
 	// handler for self target
 	win.webContents.on('will-navigate', (e, url) =>
-			url != win.webContents.getURL() &&
-			(e.preventDefault(), shell.openExternal(url))
+		url != win.webContents.getURL() &&
+		(e.preventDefault(), shell.openExternal(url))
 	);
 
 	ConfigManager.watchOption("launcher.openDevTools")(state =>
-			state ? win.webContents.openDevTools() : win.webContents.closeDevTools()
+		state ? win.webContents.openDevTools() : win.webContents.closeDevTools()
 	)
 
 });
 
-exports.createMainWindow = createMainWindow;
-
-const restoreMainWindow = () => {
-	if (win == void 0) createMainWindow();
-	else restoreWindow();
+exports.restore = () => {
+	if (win == void 0) {
+		exports.create();
+	} else {
+    if (!win.isVisible()) win.show();
+    if (win.isMinimized()) win.restore();
+    win.focus();
+	}
 };
 
-exports.restoreMainWindow = restoreMainWindow;
+exports.destroy = () => win != void 0 && win.destroy();
+exports.setProgressBar = (progress) => win != void 0 && win.setProgressBar(progress);
+exports.send = (channel, ...args) => win != void 0 && win.webContents.send(channel, ...args);
