+++ src/main/helpers.js
@@ -1,29 +1,5 @@
-const { shell, BrowserWindow, nativeImage } = require("electron");
+const { shell, nativeImage } = require("electron");
 const path = require('path');
-
-exports.windowSend = function(cmd) {
-  const windows = BrowserWindow.getAllWindows();
-  if (windows[0]) {
-    console.log(`Sending "${cmd}"`);
-    windows[0].webContents.send(cmd);
-  } else {
-    console.log(`Tried to send "${cmd}", but could not find window`);
-  }
-}
-
-exports.restoreWindow = function() {
-  const windows = BrowserWindow.getAllWindows();
-
-  if (windows[0]) {
-    const win = windows[0];
-    if (!win) return;
-    if (!win.isVisible()) win.show();
-    if (win.isMinimized()) win.restore();
-    win.focus();
-  } else {
-    console.log(`Tried to restore window, but could not find window`);
-  }
-}
 
 exports.openDir = function(path) {
   // logger.debug("[Main]", "{OpenMineDir}", "Path:", path);
